Added folder model structure for test logs I decided to move the each model to it's own folder inside models. This way we can store all test logs, tests and performance results of this model inside its corresponding directory.
