+++ Tools/parametric_model/model_results.yml
@@ -1,27 +1,27 @@
 coefficients:
-  c_d_wing_xy_lin: 0.08189859137907289
-  c_d_wing_xy_offset: 0.02817137153038889
-  c_d_wing_xy_quad: 0.06100147573497677
-  c_d_wing_xy_stall: -3.563563361754457e-13
-  c_d_wing_y_offset: -7.790947777784442
-  c_d_x_flap_lin: -0.022235685492639462
-  c_d_x_flap_offset: -0.028171371533127536
-  c_d_x_flap_quad: 0.03294684395591521
-  c_d_y_flap_1: 0.20264572827171767
-  c_d_y_flap_2: -0.19741043861644486
-  c_l_wing_xy_lin: 0.06231244447455825
-  c_l_wing_xy_offset: 0.2408172871289555
-  c_l_wing_xy_stall: -4.855366915434744e-13
-  forward_rot_drag_lin: -10.22249749162915
-  forward_rot_drag_offset: 0.13461517336983939
-  forward_rot_thrust_lin: 1.9892354946144866
-  forward_rot_thrust_offset: -0.9149083533601631
-  forward_rot_thrust_quad: 43.581510352838414
-  intercept: -0.02994931820323951
-  vert_rot_drag_lin: -0.3754098258960826
-  vert_rot_drag_offset: 0.18445079364237837
-  vert_rot_thrust_lin: 9.82976740908115
-  vert_rot_thrust_offset: -3.8946998178292693
-  vert_rot_thrust_quad: -5.039517400186501
+  c_d_wing_xy_lin: -4.730289945754952
+  c_d_wing_xy_offset: 13.429352875935605
+  c_d_wing_xy_quad: 6.20348614097491
+  c_d_wing_xy_stall: 2.3018875201376505
+  c_d_wing_y_offset: -4.423655568723647
+  c_d_x_flap_lin: 0.5135338776826988
+  c_d_x_flap_offset: 4.836376527061735
+  c_d_x_flap_quad: -0.3189330511258948
+  c_d_y_flap_1: 0.21300864520706705
+  c_d_y_flap_2: -0.2088619409353575
+  c_l_wing_xy_lin: 0.13499219150255545
+  c_l_wing_xy_offset: 0.07921145523787143
+  c_l_wing_xy_stall: 0.6466633309339371
+  forward_rot_drag_lin: 1.2797290130073364
+  forward_rot_drag_offset: 1.60238993079344
+  forward_rot_thrust_lin: -0.6739148416023818
+  forward_rot_thrust_offset: -0.008423580885009063
+  forward_rot_thrust_quad: 49.69184979689383
+  intercept: -0.016429546041228676
+  vert_rot_drag_lin: -0.05154995539494904
+  vert_rot_drag_offset: -0.11658323179078314
+  vert_rot_thrust_lin: 3.8630995900689133
+  vert_rot_thrust_offset: -3.608313705143786
+  vert_rot_thrust_quad: 5.016895583440511
 metrics:
-  R2: 0.5684415365467737
+  R2: 0.6477012891100123
